crypto/encoding: swap dependency direction
