+++ crypto/Cargo.toml
@@ -31,6 +31,9 @@ blst = { version = "0.3.12", optional = true }
 
 proptest = { version = "1.1", optional = true }
 
+tezos_data_encoding.workspace = true
+nom.workspace = true
+
 [dev-dependencies]
 serde_json = "1.0"
